Rename file pageMetadata
