+++ themes/cclw/pages/homepage.tsx
@@ -14,7 +14,7 @@ import { PoweredBy } from "@cclw/components/PoweredBy";
 import { FeatureDiscover } from "@cclw/components/FeatureDiscover";
 import { HelpUs } from "@cclw/components/HelpUs";
 import { FeatureSearch } from "@cclw/components/FeatureSearch";
-import { PAGE_DESCRIPTION, APP_NAME } from "@cclw/constants/description";
+import { PAGE_DESCRIPTION, APP_NAME } from "@cclw/constants/pageMetadata";
 import { Heading } from "@components/typography/Heading";
 
 const WorldMap = dynamic(() => import("@components/map/WorldMap"), {
