+++ themes/mcf/pages/homepage.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import Layout from "@components/layouts/LandingPage";
 
 import { Header, Footer, Hero, ClimatePolicyRadarBanner, AboutTheFunds, ContextualSearchContent, AboutClimateProjectExplorer } from "@mcf/components";
-import { PAGE_DESCRIPTION, APP_NAME } from "@mcf/constants/description";
+import { PAGE_DESCRIPTION, APP_NAME } from "@mcf/constants/pageMetadata";
 
 type TProps = {
   handleSearchInput: (term: string, filter?: string, filterValue?: string) => void;
